+++ storybook/cypress/e2e/document-card.cy.ts
@@ -64,4 +64,13 @@ describe("Document Card", () => {
 
     cy.get("dso-document-card.hydrated").matchImageSnapshot();
   });
+
+  it("should show characteristics as bright labels, next to each other", () => {
+    cy.visit("http://localhost:45000/iframe.html?id=core-document-card--with-characteristics")
+      .get("dso-document-card.hydrated")
+      .find("[slot='characteristics'] > dso-label[status='bright']")
+      .should("have.length", 2);
+
+    cy.get("dso-document-card.hydrated").matchImageSnapshot();
+  });
 });